requirements.txt
Django==4.2.16
Pillow==10.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
channels==4.1.0
channels-redis==4.2.0
redis==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Commented out cloud services - using local storage only
# boto3==1.35.24
# django-storages==1.14.4
# firebase-admin==6.1.0
# firebase input-otp==0.1.5
# azure-ai-documentintelligence==1.0.0